fix: grade system
add: elk articleId

+++ store/src/main/java/com/backend/store/dataobject/request/ChangeArticleGradeRequest.java
@@ -10,6 +10,6 @@ import javax.validation.constraints.NotNull;
 public class ChangeArticleGradeRequest {
     @NotNull
     private Long articleId;
-    private Long newGrade;
+    private Float newGrade;
     private Long customerNumber;
 }
